restructure folder and add loaders
